office_agent/requirements.txt
google-generativeai
fastapi
uvicorn